+++ android/app/src/main/java/com/yagubogu/YaguBoguApplication.kt
@@ -6,7 +6,6 @@ import com.google.firebase.crashlytics.FirebaseCrashlytics
 import com.yagubogu.common.YaguBoguDebugTree
 import com.yagubogu.common.YaguBoguReleaseTree
 import com.yagubogu.data.datasource.LocationLocalDataSource
-import com.yagubogu.data.datasource.MemberRemoteDataSource
 import com.yagubogu.data.datasource.StadiumRemoteDataSource
 import com.yagubogu.data.datasource.StatsRemoteDataSource
 import com.yagubogu.data.datasource.TalksRemoteDataSource
@@ -14,6 +13,7 @@ import com.yagubogu.data.datasource.TokenRemoteDataSource
 import com.yagubogu.data.datasource.auth.AuthRemoteDataSource
 import com.yagubogu.data.datasource.checkin.CheckInRemoteDataSource
 import com.yagubogu.data.datasource.game.GameRemoteDataSource
+import com.yagubogu.data.datasource.member.MemberRemoteDataSource
 import com.yagubogu.data.network.RetrofitInstance
 import com.yagubogu.data.network.TokenManager
 import com.yagubogu.data.repository.AuthDefaultRepository
