+++ android/app/src/main/java/com/yagubogu/data/dto/request/member/MemberFavoriteRequest.kt
@@ -1,5 +1,4 @@
-package com.yagubogu.data.dto.request
-
+package com.yagubogu.data.dto.request.member
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -7,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class MemberFavoriteRequest(
     @SerialName("teamCode")
-    val teamCode: String
+    val teamCode: String,
 )
