+++ android/app/src/main/java/com/yagubogu/data/dto/request/member/MemberNicknameRequest.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.data.dto.request
+package com.yagubogu.data.dto.request.member
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
